Initial edits (#1)
## Changes

- (Solidity) minor changes to business logic for clarity
- (Solidity) changed comments to natspec where applicable
- (Solidity) implemented erc20 to remove external open zeppelin
dependency
- (Solidity) updated license in one contract to `MIT` (to match others)
- (.DS_Store) found and removed a hidden OSX file

+++ solidity-examples/contracts/12 - Contract.sol
@@ -1,12 +1,15 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.17;
 
+/// @title Contract Information
 contract ContractInfo {
-   function getContractAddress() public view returns (address){
-        return address(this);
-   }
+     /// @notice Get contract address.
+     function getContractAddress() public view returns (address) {
+          return address(this);
+     }
 
-   function getContractBalance() public view returns (uint) {
-        return address(this).balance;
-   }
+     /// @notice Get contract balance.
+     function getContractBalance() public view returns (uint256) {
+          return address(this).balance;
+     }
 }
